2.15 Модульный JS-доработка render.js
